Make the capture gate fire at all
`referencedNames(in:)` walked a node's children and never the node itself, so a
caller handing it a bare `DeclReferenceExprSyntax` got back the empty set — the
one node it was asked about was the one node never examined.

Every caller but one passes a container (an accessor block, an argument list, a
closure), where a root that is itself a reference cannot occur. The exception is
the `toggle` check inside `requiresCapture`, which passes the base of
`isExpanded.toggle()`. So that branch has existed since the gate was written and
has never once fired: `isExpanded = true` was gated and `isExpanded.toggle()`
was reported, the same mutation under two spellings.

`assignedNames` had already hit this and worked around it in place, re-inserting
the element's own name after calling here — the codebase knew about the trap and
patched it at one of the two sites.

Found by hand-classifying `LintIssueRow`, the file this gate was calibrated
against when it was written, and noticing its expand button was still reported.

Corpus effect: 134 -> 133. One finding, and it is exactly the property predicted
by hand before the fix was written.
